refactor: centralize save loading logic
